+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.2.1
 
 - Disabled the enhanced editor toggle button by default.
+- Disabled the line break indicator by default.
 
 ## 1.2.0